Resizing an image with RabbitMQ/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncpg==0.30.0
bcrypt==4.3.0
bokeh==2.4.2
certifi==2025.4.26
charset-normalizer==2.0.12
click==8.2.1
colorama==0.4.6
crispy-bootstrap5==2025.6
Django==5.2.4
django-crispy-forms==2.4
django-crontab==0.7.1
django-environ==0.8.1
django-extensions==3.1.3
django-htmx==1.6.0
django-pwa==2.0.1
django-widget-tweaks==1.4.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.5
orjson==3.10.18
packaging==25.0
passlib==1.7.4
pika==1.3.2
pillow==11.2.1
pydantic==2.11.5
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.26.0
rich==14.0.0
rich-toolkit==0.14.7
scalar_fastapi==1.0.3
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
sqlparse==0.5.3
starlette==0.46.2
tornado==6.4.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
